pyproject.toml
[project]
name = "amazon-price-competitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.34",
    "openai>=2.1.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "tinydb>=4.8.2",
]
